Moving GC types
